ENH: Make size_type uniform accross vxl
The value type for size_type was different in
different parts of vxl.  In all cases using
an alias to size_t seems to work well.

This has the advantage of being consistent
in all parts of vxl.

+++ core/vnl/vnl_vector.h
@@ -257,7 +257,7 @@ class vnl_vector
 
   //: Type defs for iterators
   typedef T element_type;
-  typedef unsigned  size_type;
+  typedef size_t  size_type;
 
   //: Type defs for iterators
   typedef T       *iterator;
